# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agile Metrics Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all Jira projects
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects/{projectId}/members/{period}:
    get:
      operationId: getProjectMembers
      tags: [metrics]
      summary: Get team member statistics for a project
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/PeriodPath"
      responses:
        "200":
          description: Member statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberStats"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/issues/{period}:
    get:
      operationId: getProjectIssues
      tags: [metrics]
      summary: Get recent issues for a project
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/PeriodPath"
      responses:
        "200":
          description: Issues list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
...
  /projects/{projectId}/issues/{period}/csv:
    get:
      operationId: downloadProjectIssuesCsv
      tags: [metrics]
      summary: Export issues as CSV
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/PeriodPath"
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
    PeriodPath:
      name: period
      in: path
      required: true
      schema:
        type: string
        enum: ["1m", "3m", "6m"]
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Project:
      type: object
      required: [id, key, name, projectType, issueCount, doneCount, inProgressCount, visible]
      properties:
        id:
          type: string
        key:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        projectType:
          type: string
        boardType:
          type: string
          enum: [scrum, kanban, simple]
          description: Agile board methodology detected from Jira boards
        methodology:
          type: string
          description: Human-readable methodology label (Scrum or Kanban)
        avatarUrl:
          type: ["string", "null"]
        issueCount:
          type: integer
        doneCount:
          type: integer
        inProgressCount:
          type: integer
        visible:
          type: boolean
        lead:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        usingMockData:
          type: boolean
          description: Whether the data is from Jira or demo/mock data
...
    MemberStats:
      type: object
      required: [accountId, displayName, issuesResolved, storyPoints, avgCycleTime, issuesInProgress]
      properties:
        accountId:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        issuesResolved:
          type: integer
        storyPoints:
          type: number
        avgCycleTime:
          type: number
        issuesInProgress:
          type: integer
    Issue:
      type: object
      required: [id, key, summary, status, issueType, priority, createdAt]
      properties:
        id:
          type: string
        key:
          type: string
        summary:
          type: string
        status:
          type: string
        issueType:
          type: string
        priority:
          type: string
        assignee:
          type: ["string", "null"]
        storyPoints:
          type: ["number", "null"]
        createdAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
        cycleTimeDays:
          type: ["number", "null"]
...